Compressed cut

+++ src/FB/Intrinsic.h
@@ -8,10 +8,10 @@
 // Copyright (c) 2007-2011 VCAD System Research Program, RIKEN.
 // All rights reserved.
 //
-// Copyright (c) 2011-2014 Institute of Industrial Science, The University of Tokyo.
+// Copyright (c) 2011-2015 Institute of Industrial Science, The University of Tokyo.
 // All rights reserved.
 //
-// Copyright (c) 2012-2014 Advanced Institute for Computational Science, RIKEN.
+// Copyright (c) 2012-2015 Advanced Institute for Computational Science, RIKEN.
 // All rights reserved.
 //
 //##################################################################################
@@ -136,7 +136,7 @@ public:
                       REAL_TYPE* G_org,
                       const int NoMedium,
                       const MediumList* mat,
-                      float* cut,
+                      long long* cut,
                       int* bid) {};
   
   
@@ -161,7 +161,7 @@ public:
                      Control* R,
                      const int NoMedium,
                      const MediumList* mat,
-                     float* cut,
+                     long long* cut,
                      int* bid) {};
