+++ src/IP/IP_Step.h
@@ -8,10 +8,10 @@
 // Copyright (c) 2007-2011 VCAD System Research Program, RIKEN.
 // All rights reserved.
 //
-// Copyright (c) 2011-2014 Institute of Industrial Science, The University of Tokyo.
+// Copyright (c) 2011-2015 Institute of Industrial Science, The University of Tokyo.
 // All rights reserved.
 //
-// Copyright (c) 2012-2014 Advanced Institute for Computational Science, RIKEN.
+// Copyright (c) 2012-2015 Advanced Institute for Computational Science, RIKEN.
 // All rights reserved.
 //
 //##################################################################################
@@ -22,7 +22,7 @@
  * @author aics
  */
 
-#include "../FB/Intrinsic.h"
+#include "Intrinsic.h"
 #include "IP_Define.h"
 #include "FBUtility.h"
 
@@ -70,7 +70,7 @@ public:
    * @param [out]    cut      カット情報
    * @param [out]    bid      境界ID
    */
-  virtual void setup(int* bcd, Control* R, const int NoMedium, const MediumList* mat, float* cut, int* bid);
+  virtual void setup(int* bcd, Control* R, const int NoMedium, const MediumList* mat, long long* cut, int* bid);
 
 };
 #endif // _IP_STEP_H_
